moved data visualization components to a folder, attribute icons only contained used icons
