+++ src/components/data-visualization/mood-bar.tsx
